+++ server/apis/interface.go
@@ -19,20 +19,26 @@ package apis
 import "github.com/gin-gonic/gin"
 
 type Handler interface {
+	ListNamespaces(c *gin.Context)
+	GetClusterSummary(c *gin.Context)
+	CreatePipeline(c *gin.Context)
 	ListPipelines(c *gin.Context)
 	GetPipeline(c *gin.Context)
+	UpdatePipeline(c *gin.Context)
+	DeletePipeline(c *gin.Context)
+	PatchPipeline(c *gin.Context)
+	CreateInterStepBufferService(c *gin.Context)
 	ListInterStepBufferServices(c *gin.Context)
 	GetInterStepBufferService(c *gin.Context)
-	ListVertices(c *gin.Context)
-	GetVertex(c *gin.Context)
-	GetVertexMetrics(c *gin.Context)
-	ListVertexPods(c *gin.Context)
-	PodLogs(c *gin.Context)
-	ListPodsMetrics(c *gin.Context)
-	GetPodMetrics(c *gin.Context)
+	UpdateInterStepBufferService(c *gin.Context)
+	DeleteInterStepBufferService(c *gin.Context)
 	ListPipelineBuffers(c *gin.Context)
-	GetVertexBuffers(c *gin.Context)
 	GetPipelineWatermarks(c *gin.Context)
+	UpdateVertex(c *gin.Context)
+	GetVerticesMetrics(c *gin.Context)
+	ListVertexPods(c *gin.Context)
+	ListPodsMetrics(c *gin.Context)
+	PodLogs(c *gin.Context)
+	GetNamespaceEvents(c *gin.Context)
 	GetPipelineStatus(c *gin.Context)
-	ListNamespaces(c *gin.Context)
 }
