+++ server/apis/v1/response_k8s_events.go
@@ -1,4 +1,4 @@
-package v1_1
+package v1
 
 import "fmt"
