+++ server/apis/v1/response_pipeline.go
@@ -1,4 +1,4 @@
-package v1_1
+package v1
 
 import "github.com/numaproj/numaflow/pkg/apis/numaflow/v1alpha1"
